Clear the index data structures before (re)indexing an object.
Previously it was possible for the index to contain stale data from a previous
indexing of an object which caused false results in queries.

This was particularly a problem when an object was first indexed with recurrence
information and then reindex with recurrence removed. Subsequent queries would
still return hits from the previous recurrence period.